feat: implement AI-driven item generation and processing services
- Added AiItemGenerationService for generating initial items using OpenAI.
- Introduced CategoryProcessingService to categorize and tag extracted items.
- Created ContentFilteringService for filtering and assessing web pages based on relevance.
- Developed DataAggregationService for aggregating and deduplicating item data.
- Implemented ItemExtractionService for extracting relevant items from web pages.
- Added SearchQueryGenerationService for generating search queries using LLM.
- Created SourceValidationService for validating source URLs of items.
- Developed WebPageRetrievalService for retrieving web pages based on search queries.
- Updated index file to export all new services for easier access.
