+++ apps/api/src/items-generator/items-generator.module.ts
@@ -1,8 +1,26 @@
 import { Module } from '@nestjs/common';
 import { ItemsGeneratorService } from './items-generator.service';
+import { AiItemGenerationService } from './steps/ai-item-generation.service';
+import { SearchQueryGenerationService } from './steps/search-query-generation.service';
+import { WebPageRetrievalService } from './steps/web-page-retrieval.service';
+import { ContentFilteringService } from './steps/content-filtering.service';
+import { ItemExtractionService } from './steps/item-extraction.service';
+import { SourceValidationService } from './steps/source-validation.service';
+import { DataAggregationService } from './steps/data-aggregation.service';
+import { CategoryProcessingService } from './steps/category-processing.service';
 
 @Module({
-  providers: [ItemsGeneratorService],
+  providers: [
+    ItemsGeneratorService,
+    AiItemGenerationService,
+    SearchQueryGenerationService,
+    WebPageRetrievalService,
+    ContentFilteringService,
+    ItemExtractionService,
+    SourceValidationService,
+    DataAggregationService,
+    CategoryProcessingService,
+  ],
   exports: [ItemsGeneratorService],
 })
 export class ItemsGeneratorModule {}
